requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-google-genai>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
streamlit>=1.40.0
plotly>=5.24.0
loguru>=0.7.2
pytest>=8.3.0
pytest-asyncio>=0.24.0
pypdf>=5.1.0
google-generativeai>=0.8.0
fpdf2>=2.8.2
